feat: add hash
